@@ -6,7 +6,7 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2",
+        "php": ">=7.4",
         "ext-json": "*",
         "composer/composer": "1.10.23 as 2.0.13",
         "composer/installers": "^1.9",
@@ -82,6 +82,7 @@
         "drupal/views_bulk_operations": "^3.10",
         "drupal/views_data_export": "^1.0",
         "drupal/views_data_export_phpspreadsheet": "^1.2",
+        "drupal/workflow_buttons": "^1.0@beta",
         "drush/drush": "^10.3",
         "ethosce/group_purl": "2.x-dev",
         "lramarojaona/purl": "^1.0@alpha",
