@@ -15,7 +15,7 @@ final class Topics {
 
   const TERM_TOPICS_ID_FIELD_USER_SOLR = 'sm_user_profile_field_vocab_topic_interest_array';
 
-  const TERM_TOPICS_ID_FIELD_GROUP_SOLR = 'ss_group_topic_name';
+  const TERM_TOPICS_ID_FIELD_GROUP_SOLR = 'sm_group_topic_name';
 
   const CONTENT_TYPE_ID_FIELD_SOLR = 'ss_global_content_type';
 
