@@ -5,6 +5,9 @@ import getTranslation from "../../../../Services/Translations";
 import Actions from "./PopupComponents/Actions";
 import PopupLayout from "./PopupComponents/Layout";
 import Author from "./components/Author";
+import TaggedUsers from "./TaggedUsers";
+import TaggedUsersModal from "./PopupComponents/TaggedUsersModal";
+
 
 class PopupForm extends React.Component {
   constructor(props) {
@@ -12,11 +15,14 @@ class PopupForm extends React.Component {
 
     this.state = {
       text: props.comment.text || "",
+      taggedUsers: this.props.comment.tagged_users || [],
     };
+    this.settings = window.drupalSettings.overview;
 
     this.textarea = React.createRef()
 
     this.submit = this.submit.bind(this);
+    this.updateTaggedUsers = this.updateTaggedUsers.bind(this);
   }
 
   submit(e, closeFunction) {
@@ -25,8 +31,15 @@ class PopupForm extends React.Component {
       e,
       this.props.comment.comment_id,
       this.state.text,
-      this.props.parentId
+      this.props.parentId,
+      this.state.taggedUsers
     );
+  }
+
+  updateTaggedUsers(users) {
+    this.setState({
+      taggedUsers: users,
+    })
   }
 
   render() {
@@ -72,20 +85,24 @@ class PopupForm extends React.Component {
                                 {this.props.title}
                               </label>
                               {this.props.showTextArea && (
-                                <div className="ecl-comment-form__textarea-wrapper">
-                                  <textarea
-                                    ref={this.textarea}
-                                    onChange={(e) =>
-                                      this.setState({ text: e.target.value })
-                                    }
-                                    value={this.state.text}
-                                    className="ecl-text-area ecl-comment-form__textarea"
-                                    id="ecl-comment-form-reply"
-                                    name=""
-                                    placeholder={getTranslation(
-                                      "comment_placeholder"
-                                    )}
-                                  />
+                                <div>
+                                  <div className="ecl-comment-form__textarea-wrapper">
+                                    <textarea
+                                      ref={this.textarea}
+                                      onChange={(e) =>
+                                        this.setState({ text: e.target.value })
+                                      }
+                                      value={this.state.text}
+                                      className="ecl-text-area ecl-comment-form__textarea"
+                                      id="ecl-comment-form-reply"
+                                      name=""
+                                      placeholder={getTranslation(
+                                        "comment_placeholder"
+                                      )}
+                                    />
+                                  </div>
+                                  <TaggedUsers taggedUsers={this.state.taggedUsers}/>
+                                  <TaggedUsersModal settings={this.settings} taggedUsers={this.state.taggedUsers} updateTaggedUsers={this.updateTaggedUsers} />
                                 </div>
                               )}
                             </div>
