@@ -25,11 +25,12 @@ const TaggedUsersModal = ({ settings, taggedUsers, updateTaggedUsers }) => {
           <Popup
             trigger={
               <span
+                className="ecl-comment-form__tagged-user"
                 style={{ cursor: "pointer" }}
                 dangerouslySetInnerHTML={{
                   __html: svg(
                     "user",
-                    "ecl-icon ecl-icon--s ecl-timestamp__icon"
+                    "ecl-icon ecl-icon--s"
                   ),
                 }}
               />
