@@ -6,8 +6,8 @@ import initCollapse from "../../../Services/CollapseOptions";
 import CommentForm from "./Partials/CommentForm";
 import LoadMore from "./Partials/LoadMore";
 import TopAnswer from "./Partials/TopAnswer";
-import {store} from 'react-notifications-component';
-import {getPrefixDiscussionEndpoint} from "./Services/UrlHelper";
+import { store } from 'react-notifications-component';
+import { getPrefixDiscussionEndpoint } from "./Services/UrlHelper";
 
 const svg = require('../../../svg/svg')
 
@@ -29,6 +29,7 @@ class CommentsDiscussion extends React.Component {
 
     this.wrapperRef = React.createRef();
 
+
     this.fetchComments = this.fetchComments.bind(this);
     this.addComment = this.addComment.bind(this);
     this.handleClickOutside = this.handleClickOutside.bind(this);
@@ -47,6 +48,9 @@ class CommentsDiscussion extends React.Component {
   componentDidMount() {
     this.fetchComments();
     document.addEventListener('click', this.handleClickOutside);
+    if (document.getElementById('ecl-comment-overview__wrapper') && this.props.highlightedCommentId != 0) {
+      document.getElementById('ecl-comment-overview__wrapper').scrollIntoView()
+    }
   }
 
   componentWillUnmount() {
@@ -113,11 +117,11 @@ class CommentsDiscussion extends React.Component {
     })
 
     axios.post(getPrefixDiscussionEndpoint() + '/' + this.props.discussionId + '/comment', {
-        text: parentId ? this.state.commentTextReply : this.state.commentText,
-        parentId: parentId,
-        taggedUsers: taggedUsers,
-      },
-      {withCredentials: true})
+      text: parentId ? this.state.commentTextReply : this.state.commentText,
+      parentId: parentId,
+      taggedUsers: taggedUsers,
+    },
+      { withCredentials: true })
       .then(function (response) {
         self.setState({
           activeAnswerForm: 0,
@@ -143,7 +147,7 @@ class CommentsDiscussion extends React.Component {
     })
 
     axios.delete(getPrefixDiscussionEndpoint() + '/' + window.drupalSettings.overview.group_id + '/' + this.props.discussionId + '/delete/' + commentId,
-      {withCredentials: true})
+      { withCredentials: true })
       .then(function (response) {
         self.addNotification('Comment deleted')
         this.setState({
@@ -205,8 +209,8 @@ class CommentsDiscussion extends React.Component {
       loading: true
     })
 
-    axios.post(getPrefixDiscussionEndpoint() + '/' +  this.props.discussionId + '/' + flagId + '/' + commentId + '/' + type,
-      {text:text, withCredentials: true})
+    axios.post(getPrefixDiscussionEndpoint() + '/' + this.props.discussionId + '/' + flagId + '/' + commentId + '/' + type,
+      { text: text, withCredentials: true })
       .then(function (response) {
         self.addNotification(`Comment ${flagId} ${type}`)
         self.fetchComments();
@@ -226,12 +230,12 @@ class CommentsDiscussion extends React.Component {
     })
 
     axios.post(`${window.drupalSettings.path.baseUrl}api/request/${request_type}/${entity_type}/${entity_id}`,
-      {reason:reason, withCredentials: true})
+      { reason: reason, withCredentials: true })
       .then(response => {
-        if(response.data && response.data.result) {
+        if (response.data && response.data.result) {
           self.addNotification(`Your request has been sent!`)
           self.fetchComments();
-        }else {
+        } else {
           self.addNotification(response.data.message, 'danger')
         }
       })
@@ -244,7 +248,7 @@ class CommentsDiscussion extends React.Component {
     })
   }
 
-  updateComment(e, commentId, text, parentId = null) {
+  updateComment(e, commentId, text, parentId = null, taggedUsers=[]) {
     e.preventDefault();
     const self = this;
     this.setState({
@@ -252,7 +256,7 @@ class CommentsDiscussion extends React.Component {
     })
 
     axios.patch(getPrefixDiscussionEndpoint() + '/' + this.props.discussionId + '/edit/' + commentId,
-      {'text': text, withCredentials: true})
+      { 'text': text, taggedUsers: taggedUsers, withCredentials: true })
       .then(function (response) {
         self.addNotification('Comment edited')
         this.setState({
@@ -293,12 +297,12 @@ class CommentsDiscussion extends React.Component {
       {
         this.state.loading &&
         <CircularProgress
-          style={{top: '50%', left: '50%', position: 'fixed', zIndex: '51'}}
+          style={{ top: '50%', left: '50%', position: 'fixed', zIndex: '51' }}
           className="loader"
           size={50}
         />
       }
-      <div ref={this.wrapperRef} className="ecl-comment-overview__items">
+      <div ref={this.wrapperRef} id='ecl-comment-overview__wrapper' className="ecl-comment-overview__items">
         {this.state.comments.map(comment => {
           return <TopAnswer
             key={comment.comment_id}
@@ -323,7 +327,7 @@ class CommentsDiscussion extends React.Component {
           />
         })}
       </div>
-      {this.state.total > this.state.totalLoaded && <LoadMore updatePage={this.updatePage}/>}
+      {this.state.total > this.state.totalLoaded && <LoadMore updatePage={this.updatePage} />}
     </div>
   }
 
