@@ -9,6 +9,7 @@ class DateField extends React.Component {
     super(props);
 
     this.handleChange = this.handleChange.bind(this);
+    const translations = window.drupalSettings.translations;
   }
 
   handleChange(e) {
@@ -19,7 +20,7 @@ class DateField extends React.Component {
     return (
       <div className="ecl-filter-sidebar__item">
         <span className="ecl-filter-sidebar__item-label" tabIndex="-1">
-          Dates
+          {this.props.translations.date_filter_label}
         </span>
         <DatePicker
           selectsRange={true}
