@@ -22,7 +22,7 @@ class SearchField extends React.Component {
         {
           this.props.showLabel &&
           <span className="ecl-filter-sidebar__item-label" tabIndex="-1">
-            {this.props.translations.search_text}
+            {this.search_label}
           </span>
         }
 
@@ -30,7 +30,7 @@ class SearchField extends React.Component {
           <div className="ecl-filter-sidebar__item-fields">
             <div className="ecl-filter-sidebar__item-field">
               <div className="ecl-form-group">
-                <input placeholder={this.search_label} value={this.props.searchText} onChange={e => this.handleChange(e)} className="ecl-text-input ecl-text-input--m" type="text"/>
+                <input placeholder={this.props.translations.search_placeholder} value={this.props.searchText} onChange={e => this.handleChange(e)} className="ecl-text-input ecl-text-input--m" type="text"/>
                 <button
                   type='submit'
                   className="ecl-search__icon"
