@@ -1,7 +1,7 @@
 import React from 'react';
 
 const svg = require('../../../../../svg/svg')
-import {timeDifferenceFromNow} from "../../../../../Services/TimeHelper";
+import { timeDifferenceFromNow } from "../../../../../Services/TimeHelper";
 import UserImage from "../../../../Utils/UserImage";
 
 class GroupResultItem extends React.Component {
@@ -17,7 +17,7 @@ class GroupResultItem extends React.Component {
     let groupVisibilityLabel = '';
     let groupClass = 'public';
     const moderationState = this.props.result.ss_group_moderation_state;
-    const groupImageFallback = '<use xlink:href="themes/custom/eic_community/dist/images/sprite/custom/sprites/custom.svg#custom--group_circle"></use>';
+    const groupImageFallback = this.props.type === 'organisation' ? 'organisation_circle' : 'group_circle';
 
     if (moderationState !== 'published') {
       groupVisibilityLabel = this.props.translations[moderationState];
@@ -33,14 +33,13 @@ class GroupResultItem extends React.Component {
           <a href={this.props.result.ss_url}>
             <figure className="ecl-teaser__image-wrapper">
               {this.props.result.ss_group_teaser_url_string !== undefined &&
-              <img className="ecl-teaser__image"
-                   src={this.props.result.ss_group_teaser_url_string}
-                   alt={this.props.result.tm_global_title}/>
+                <img className="ecl-teaser__image"
+                  src={this.props.result.ss_group_teaser_url_string}
+                  alt={this.props.result.tm_global_title} />
               }
               {this.props.result.ss_group_teaser_url_string === undefined &&
-              <div className="ecl-teaser__image-fallback-wrapper">
-                <svg className="ecl-icon ecl-icon--3xl" focusable="false" aria-hidden="true" dangerouslySetInnerHTML={{__html: groupImageFallback }} />
-              </div>
+                <div className="ecl-teaser__image-fallback-wrapper" dangerouslySetInnerHTML={{ __html: svg(groupImageFallback, 'ecl-icon ecl-icon--3xl') }}>
+                </div>
               }
             </figure>
           </a>
@@ -61,7 +60,7 @@ class GroupResultItem extends React.Component {
                     <span className="ecl-author__label ">
                       {!this.props.isAnonymous ? (
                         <a href={this.props.result.ss_global_user_url}
-                            className="ecl-link ecl-link--standalone ecl-author__link">
+                          className="ecl-link ecl-link--standalone ecl-author__link">
                           {this.props.result.ss_global_fullname}
                         </a>
                       ) : this.props.result.ss_global_fullname}
@@ -82,32 +81,32 @@ class GroupResultItem extends React.Component {
 
                 <div className="ecl-timestamp ">
                   <span
-                    dangerouslySetInnerHTML={{__html: svg('time', 'ecl-icon ecl-icon--s ecl-timestamp__icon')}}
+                    dangerouslySetInnerHTML={{ __html: svg('time', 'ecl-icon ecl-icon--s ecl-timestamp__icon') }}
                   />
                   <time className="ecl-timestamp__label">Last
-                    activity {timeDifferenceFromNow(this.props.result.ss_drupal_timestamp)}</time>
+                    activity {timeDifferenceFromNow(this.props.result.ss_drupal_changed_timestamp)}</time>
                 </div>
               </div>
             </div>
             <div className="ecl-teaser__meta-footer">
               <div className="ecl-teaser__stats">
                 <div className="ecl-teaser__stat">
-                  <div dangerouslySetInnerHTML={{__html: svg('group', 'ecl-icon--xs ecl-teaser__stat-icon')}}/>
+                  <div dangerouslySetInnerHTML={{ __html: svg('group', 'ecl-icon--xs ecl-teaser__stat-icon') }} />
                   <span className="ecl-teaser__stat-label">{this.props.translations.members}</span>
                   <span className="ecl-teaser__stat-value">{this.props.result.its_group_statistic_members || 0}</span>
                 </div>
                 <div className="ecl-teaser__stat">
-                  <div dangerouslySetInnerHTML={{__html: svg('comment', 'ecl-icon--xs ecl-teaser__stat-icon')}}/>
+                  <div dangerouslySetInnerHTML={{ __html: svg('comment', 'ecl-icon--xs ecl-teaser__stat-icon') }} />
                   <span className="ecl-teaser__stat-label">{this.props.translations.reactions}</span>
                   <span className="ecl-teaser__stat-value">{this.props.result.its_group_statistic_comments || 0}</span>
                 </div>
                 <div className="ecl-teaser__stat">
-                  <div dangerouslySetInnerHTML={{__html: svg('documents', 'ecl-icon--xs ecl-teaser__stat-icon')}}/>
+                  <div dangerouslySetInnerHTML={{ __html: svg('documents', 'ecl-icon--xs ecl-teaser__stat-icon') }} />
                   <span className="ecl-teaser__stat-label">{this.props.translations.documents}</span>
                   <span className="ecl-teaser__stat-value">{this.props.result.its_group_statistic_files || 0}</span>
                 </div>
                 <div className="ecl-teaser__stat">
-                  <div dangerouslySetInnerHTML={{__html: svg('like', 'ecl-icon--xs ecl-teaser__stat-icon')}}/>
+                  <div dangerouslySetInnerHTML={{ __html: svg('like', 'ecl-icon--xs ecl-teaser__stat-icon') }} />
                   <span className="ecl-teaser__stat-label">{this.props.translations.like}</span>
                   <span className="ecl-teaser__stat-value">{this.props.result.its_flag_recommend_group || 0}</span>
                 </div>
