@@ -1,4 +1,4 @@
-import {event_infos_items} from "../../data/event.data";
+import { event_infos_items } from "../../data/event.data";
 
 export default {
   title: 'Bundles / Group',
@@ -11,7 +11,6 @@ import AboutFullTemplate from '@theme/patterns/compositions/group/group.full--ab
 import AttachmentTemplate from '@theme/patterns/components/attachment.html.twig';
 import DiscussionFullTemplate from '@theme/patterns/compositions/group/group.full--discussion.html.twig';
 import DiscussionThreadTemplate from '@theme/patterns/compositions/discussion-thread.html.twig';
-import EventTeaserTemplate from '@theme/patterns/compositions/event/event.teaser.html.twig';
 import ExtendedListTemplate from '@theme/patterns/compositions/extended-list/extended-list.html.twig';
 import FileListFullTemplate from '@theme/patterns/compositions/filelist/filelist.full.html.twig';
 import FilelistTemplate from '@theme/patterns/components/filelist.html.twig';
@@ -21,7 +20,6 @@ import FrontpageTemplate from '@theme/patterns/compositions/group/group.full--fr
 import OverviewTemplate from '@theme/patterns/compositions/group/group.overview.html.twig';
 import SimpleBannerTemplate from '@theme/patterns/compositions/simple-banner.html.twig';
 import StateBannerTemplate from '@theme/patterns/compositions/state-banner.html.twig';
-import StoryTeaserTemplate from '@theme/patterns/compositions/story/story.teaser.html.twig';
 import TeaserTemplate from '@theme/patterns/compositions/group/group.teaser.html.twig';
 import TeaserOverviewTemplate from '@theme/patterns/compositions/teaser-overview.html.twig';
 import WikiFullTemplate from '@theme/patterns/compositions/group/group.full--wiki.html.twig';
@@ -38,7 +36,6 @@ import EventTemplate from '@theme/patterns/components/event.html.twig';
 
 import activity_stream from '@theme/data/activity-stream.data';
 import attachment from '@theme/data/attachment.data';
-import author from '@theme/data/author.data';
 import breadcrumb from '@theme/data/breadcrumb.data';
 import common from '@theme/data/common.data';
 import comments from '@theme/data/comments.data';
@@ -1821,7 +1818,7 @@ export const GroupEventDetailed = () => GroupEventFullTemplate(
         title: 'Organised by',
         content: 'Organiser 1, Organiser 2, Organiser 3'
       },
-      action:{
+      action: {
         url: 'http;//google.be',
         label: 'Sign up'
       }
