@@ -21,7 +21,7 @@ import teaserOverview from '@theme/data/teaser-overview.data';
 import topMenu from '@theme/data/top-menu.data';
 
 import { mockItems } from '@theme/snippets';
-import {without} from "../../snippets";
+import { without } from "../../snippets";
 
 const overview = {
   breadcrumb: breadcrumb,
@@ -240,6 +240,7 @@ export const MyActivity = () =>
     {
       breadcrumb: breadcrumb,
       common: common,
+      icon_file_path: common.icon_file_path,
       mainmenu: mainmenu,
       searchform: searchform,
       subnavigation: subnavigation.myActivity,
