@@ -8,9 +8,9 @@ import { without } from '@theme/snippets';
 
 export const Base = () => BlockquoteTemplate(without(blockquote, 'author', 'image'));
 
-export const Author = () => BlockquoteTemplate(without(blockquote, 'image'));
+export const Author = () => BlockquoteTemplate({...without(blockquote, 'image'), author: true});
 
-export const AuthorAndImage = () => BlockquoteTemplate(blockquote);
+export const AuthorAndImage = () => BlockquoteTemplate({...blockquote, author: true});
 
 export default {
   title: 'Components / Blockquote',
